requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.57.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.24
langchain-core==0.3.56
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langsmith==0.3.38
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
tenacity==9.1.2
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
zstandard==0.23.0
